fix changes for pr 3
