--- screens/NavigationBar.js
@@ -1,72 +1,0 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
-import { Dimensions } from 'react-native';
-
-import Navigator from 'native-navigation';
-
-import Screen from '../components/Screen';
-import Row from '../components/Row';
-
-const propTypes = {};
-const defaultProps = {};
-const contextTypes = {
-  nativeNavigationInstanceId: PropTypes.string,
-};
-
-const { width } = Dimensions.get('window');
-
-export default class NavigationExampleScreen extends Component {
-  state = {};
-
-  render() {
-    return (
-      <Screen>
-        <Navigator.Config {...this.state} />
-        <Row
-          title="Title"
-          onPress={() => this.setState({
-            title: 'A title',
-            subtitle: undefined,
-            rightButtons: undefined,
-          })}
-        />
-        <Row
-          title="Title and subtitle"
-          onPress={() => this.setState({
-            title: 'A title',
-            subtitle: 'A subtitle',
-            rightButtons: undefined,
-          })}
-        />
-        <Row
-          title="Right button with title"
-          onPress={() => this.setState({
-            title: 'A title',
-            subtitle: undefined,
-            rightButtons: [{ title: 'Hello' }],
-          })}
-        />
-        <Row
-          title="Right button with system item"
-          onPress={() => this.setState({
-            title: 'A title',
-            subtitle: undefined,
-            rightButtons: [{ systemItem: 'add' }],
-          })}
-        />
-        <Row
-          title="Several right buttons"
-          onPress={() => this.setState({
-            title: 'A title',
-            subtitle: undefined,
-            rightButtons: [{ systemItem: 'add' }, { systemItem: 'edit' }],
-          })}
-        />
-      </Screen>
-    );
-  }
-}
-
-NavigationExampleScreen.defaultProps = defaultProps;
-NavigationExampleScreen.propTypes = propTypes;
-NavigationExampleScreen.contextTypes = contextTypes;
